rename and delete unuseful validate method
