+++ auth-service/src/main/java/com/wwlei/authservice/repo/jpa/PermissionRepository.java
@@ -1,4 +1,4 @@
-package com.wwlei.authservice.repo.jpaRepo;
+package com.wwlei.authservice.repo.jpa;
 
 import com.wwlei.authservice.repo.model.Permission;
 import org.springframework.data.jpa.repository.JpaRepository;
